[INDIDetector->[SensorInterface/Spectrograph/Detector]] Created SensorInterface and Spectrograph, Detector subclasses. TODO: work in indi-3rdparty to conform with this new standard. The sensor interface more represents the mean of single-dimensional blob devices. The first implemented drivers will be spectrographs and radio receivers as spectrograph since sampling time can represent spectra in both cases and continuum can be done both in 1element samples and chunks or even spectra. The BLOB management is done in the base interface class. Subclasses should implement properties or further BLOBs by adding methods or functionalities into themselves. Helper methods also should be present into child classes only.
