+++ CMakeLists.txt
@@ -400,7 +400,9 @@ SET(indidriver_CXX_SRC
     ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indiproperty.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indiccd.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indiccdchip.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indisensorinterface.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indidetector.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indispectrograph.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/inditelescope.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indifilterwheel.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indifocuserinterface.cpp
@@ -773,12 +775,12 @@ install(TARGETS indi_script_telescope RUNTIME DESTINATION bin)
 
 ################ Detector Simulator ################
 
-SET(detectorsimulator_SRC
-    ${CMAKE_CURRENT_SOURCE_DIR}/drivers/detector/detector_simulator.cpp)
+SET(spectrographsimulator_SRC
+    ${CMAKE_CURRENT_SOURCE_DIR}/drivers/spectrograph/spectrograph_simulator.cpp)
 
-add_executable(indi_simulator_detector ${detectorsimulator_SRC})
-target_link_libraries(indi_simulator_detector indidriver)
-install(TARGETS indi_simulator_detector RUNTIME DESTINATION bin)
+add_executable(indi_simulator_spectrograph ${spectrographsimulator_SRC})
+target_link_libraries(indi_simulator_spectrograph indidriver)
+install(TARGETS indi_simulator_spectrograph RUNTIME DESTINATION bin)
 
 ########### CCD Simulator ##############
 SET(ccdsimulator_SRC
@@ -1760,7 +1762,9 @@ if (INDI_BUILD_DRIVERS OR INDI_BUILD_CLIENT OR INDI_BUILD_QT5_CLIENT)
         ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/defaultdevice.h
         ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indiccd.h
         ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indiccdchip.h
-        ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indidetector.h
+        ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indisensorinterface.h
+        ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indispectrograph.h
+        ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indireceiver.h
         ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indifilterwheel.h
         ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indifocuserinterface.h
         ${CMAKE_CURRENT_SOURCE_DIR}/libs/indibase/indiweatherinterface.h
