+++ drivers/spectrograph/spectrograph_simulator.cpp
@@ -17,7 +17,7 @@
     Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 */
 
-#include "detector_simulator.h"
+#include "spectrograph_simulator.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
@@ -88,8 +88,8 @@ RadioSim::~RadioSim()
 ***************************************************************************************/
 bool RadioSim::Connect()
 {
-    LOG_INFO("Simulator Detector connected successfully!");
-    // Let's set a timer that checks teleDetectors status every POLLMS milliseconds.
+    LOG_INFO("Simulator Spectrograph connected successfully!");
+    // Let's set a timer that checks teleSpectrographs status every POLLMS milliseconds.
     // JM 2017-07-31 SetTimer already called in updateProperties(). Just call it once
     //SetTimer(POLLMS);
     streamPredicate = 0;
@@ -105,15 +105,14 @@ bool RadioSim::Connect()
 ***************************************************************************************/
 bool RadioSim::Disconnect()
 {
-    InCapture = false;
-    PrimaryDetector.setContinuumBufferSize(1);
-    PrimaryDetector.setSpectrumBufferSize(1);
+    InIntegration = false;
+    setBufferSize(1);
     pthread_mutex_lock(&condMutex);
     streamPredicate = 1;
     terminateThread = true;
     pthread_cond_signal(&cv);
     pthread_mutex_unlock(&condMutex);
-    LOG_INFO("Simulator Detector disconnected successfully!");
+    LOG_INFO("Simulator Spectrograph disconnected successfully!");
     return true;
 }
 
@@ -122,7 +121,7 @@ bool RadioSim::Disconnect()
 ***************************************************************************************/
 const char *RadioSim::getDefaultName()
 {
-    return "Detector Simulator";
+    return "Spectrograph Simulator";
 }
 
 /**************************************************************************************
@@ -130,20 +129,20 @@ const char *RadioSim::getDefaultName()
 ***************************************************************************************/
 bool RadioSim::initProperties()
 {
-    // We set the Detector capabilities
-    uint32_t cap = DETECTOR_CAN_ABORT | DETECTOR_HAS_CONTINUUM | DETECTOR_HAS_SPECTRUM | DETECTOR_HAS_STREAMING;
-    SetDetectorCapability(cap);
+    // We set the Spectrograph capabilities
+    uint32_t cap = SENSOR_CAN_ABORT | SENSOR_HAS_STREAMING;
+    SetSpectrographCapability(cap);
 
     // Must init parent properties first!
-    INDI::Detector::initProperties();
+    INDI::Spectrograph::initProperties();
 
-    PrimaryDetector.setMinMaxStep("DETECTOR_CAPTURE", "DETECTOR_CAPTURE_VALUE", 0.001, 86164.092, 0.001, false);
-    PrimaryDetector.setMinMaxStep("DETECTOR_SETTINGS", "DETECTOR_FREQUENCY", 2.4e+7, 2.0e+9, 1, false);
-    PrimaryDetector.setMinMaxStep("DETECTOR_SETTINGS", "DETECTOR_SAMPLERATE", 1.0e+6, 2.0e+6, 1, false);
-    PrimaryDetector.setMinMaxStep("DETECTOR_SETTINGS", "DETECTOR_GAIN", 0.0, 25.0, 0.1, false);
-    PrimaryDetector.setMinMaxStep("DETECTOR_SETTINGS", "DETECTOR_BANDWIDTH", 0, 0, 0, false);
-    PrimaryDetector.setMinMaxStep("DETECTOR_SETTINGS", "DETECTOR_BITSPERSAMPLE", 16, 16, 0, false);
-    PrimaryDetector.setCaptureExtension("fits");
+    setMinMaxStep("DETECTOR_CAPTURE", "DETECTOR_CAPTURE_VALUE", 0.001, 86164.092, 0.001, false);
+    setMinMaxStep("DETECTOR_SETTINGS", "DETECTOR_FREQUENCY", 2.4e+7, 2.0e+9, 1, false);
+    setMinMaxStep("DETECTOR_SETTINGS", "DETECTOR_SAMPLERATE", 1.0e+6, 2.0e+6, 1, false);
+    setMinMaxStep("DETECTOR_SETTINGS", "DETECTOR_GAIN", 0.0, 25.0, 0.1, false);
+    setMinMaxStep("DETECTOR_SETTINGS", "DETECTOR_BANDWIDTH", 0, 0, 0, false);
+    setMinMaxStep("DETECTOR_SETTINGS", "DETECTOR_BITSPERSAMPLE", 16, 16, 0, false);
+    setIntegrationFileExtension("fits");
 
     // Add Debug, Simulator, and Configuration controls
     addAuxControls();
@@ -158,52 +157,46 @@ bool RadioSim::initProperties()
 *********************************************************************************************/
 bool RadioSim::updateProperties()
 {
-    // Call parent update properties first
-    INDI::Detector::updateProperties();
-
     if (isConnected())
     {
-        // Let's get parameters now from Detector
+        // Let's get parameters now from Spectrograph
         setupParams();
 
         // Start the timer
         SetTimer(POLLMS);
     }
 
-    return true;
+    return INDI::Spectrograph::updateProperties();
 }
 
 /**************************************************************************************
-** Setting up Detector parameters
+** Setting up Spectrograph parameters
 ***************************************************************************************/
 void RadioSim::setupParams()
 {
-    // Our Detector is an 8 bit Detector, 100MHz frequency 1MHz bandwidth.
-    SetDetectorParams(1000000.0, 100000000.0, 16, 0.0, 25.0);
+    // Our Spectrograph is an 8 bit Spectrograph, 100MHz frequency 1MHz bandwidth.
+    setParams(1000000.0, 100000000.0, 16, 0.0, 25.0);
 }
 
 /**************************************************************************************
 ** Client is asking us to start an exposure
 ***************************************************************************************/
-bool RadioSim::StartCapture(float duration)
+bool RadioSim::StartIntegration(float duration)
 {
-    CaptureRequest = duration;
-    AbortCapture();
+    IntegrationRequest = duration;
+    AbortIntegration();
 
-    // Since we have only have one Detector with one chip, we set the exposure duration of the primary Detector
-    PrimaryDetector.setCaptureDuration(duration);
-    int to_read = PrimaryDetector.getSampleRate() * PrimaryDetector.getCaptureDuration() * abs(PrimaryDetector.getBPS()) / 8;
+    // Since we have only have one Spectrograph with one chip, we set the exposure duration of the primary Spectrograph
+    setIntegrationTime(duration);
+    int to_read = getSampleRate() * getIntegrationTime() * abs(getBPS()) / 8;
 
-    PrimaryDetector.setContinuumBufferSize(to_read);
-    if(HasSpectrum()) {
-        PrimaryDetector.setSpectrumBufferSize(SPECTRUM_SIZE * abs(PrimaryDetector.getBPS()) / 8);
-    }
-    InCapture = true;
+    setBufferSize(to_read);
+    InIntegration = true;
 
     gettimeofday(&CapStart, nullptr);
     if(HasStreaming()) {
-        Streamer->setPixelFormat(INDI_MONO, PrimaryDetector.getBPS());
-        Streamer->setSize(PrimaryDetector.getContinuumBufferSize() * 8 / abs(PrimaryDetector.getBPS()), 1);
+        Streamer->setPixelFormat(INDI_MONO, getBPS());
+        Streamer->setSize(getBufferSize() * 8 / abs(getBPS()), 1);
     }
 
     // We're done
@@ -213,15 +206,15 @@ bool RadioSim::StartCapture(float duration)
 /**************************************************************************************
 ** Client is updating capture settings
 ***************************************************************************************/
-bool RadioSim::CaptureParamsUpdated(float sr, float freq, float bps, float bw, float gain)
+bool RadioSim::paramsUpdated(float sr, float freq, float bps, float bw, float gain)
 {
     INDI_UNUSED(gain);
     INDI_UNUSED(freq);
     INDI_UNUSED(bps);
     INDI_UNUSED(bw);
-    PrimaryDetector.setBPS(16);
-    PrimaryDetector.setBandwidth(100000);
-    PrimaryDetector.setSampleRate(sr);
+    setBPS(16);
+    setBandwidth(100000);
+    setSampleRate(sr);
 
     return true;
 }
@@ -229,10 +222,10 @@ bool RadioSim::CaptureParamsUpdated(float sr, float freq, float bps, float bw, f
 /**************************************************************************************
 ** Client is asking us to abort a capture
 ***************************************************************************************/
-bool RadioSim::AbortCapture()
+bool RadioSim::AbortIntegration()
 {
-    if(InCapture) {
-        InCapture = false;
+    if(InIntegration) {
+        InIntegration = false;
     }
     return true;
 }
@@ -251,7 +244,7 @@ float RadioSim::CalcTimeLeft()
                 (double)(CapStart.tv_sec * 1000.0 + CapStart.tv_usec / 1000);
     timesince = timesince / 1000;
 
-    timeleft = CaptureRequest - timesince;
+    timeleft = IntegrationRequest - timesince;
     return timeleft;
 }
 
@@ -265,19 +258,19 @@ void RadioSim::TimerHit()
     if (isConnected() == false)
         return; //  No need to reset timer if we are not connected anymore
 
-    if (InCapture)
+    if (InIntegration)
     {
         timeleft = CalcTimeLeft();
         if(timeleft <= 0.0)
         {
             /* We're done capturing */
-            LOG_INFO("Capture done, expecting data...");
+            LOG_INFO("Integration done, expecting data...");
             timeleft = 0.0;
             grabData();
         }
 
-        // This is an over simplified timing method, check DetectorSimulator and RadioSimDetector for better timing checks
-        PrimaryDetector.setCaptureLeft(timeleft);
+        // This is an over simplified timing method, check SpectrographSimulator and RadioSimSpectrograph for better timing checks
+        setIntegrationLeft(timeleft);
     }
 
     SetTimer(POLLMS);
@@ -289,28 +282,20 @@ void RadioSim::TimerHit()
 ***************************************************************************************/
 void RadioSim::grabData()
 {
-    if(InCapture) {
+    if(InIntegration) {
 
         LOG_INFO("Downloading...");
-        InCapture = false;
+        InIntegration = false;
 
         uint8_t* continuum;
-        int size = PrimaryDetector.getContinuumBufferSize() * 8 / abs(PrimaryDetector.getBPS());
+        int size = getBufferSize() * 8 / abs(getBPS());
 
         //Fill the continuum
-        if(HasContinuum()) {
-            continuum = PrimaryDetector.getContinuumBuffer();
-            WhiteNoise(continuum, size, PrimaryDetector.getBPS());
-
-            //Create the spectrum
-            if(HasSpectrum()) {
-                uint8_t* spectrum = PrimaryDetector.getSpectrumBuffer();
-                Spectrum(continuum, spectrum, size, SPECTRUM_SIZE, PrimaryDetector.getBPS());
-            }
-        }
+        continuum = getBuffer();
+        WhiteNoise(continuum, size, getBPS());
 
         LOG_INFO("Download complete.");
-        CaptureComplete(&PrimaryDetector);
+        IntegrationComplete();
     }
 }
 
@@ -353,7 +338,7 @@ void * RadioSim::streamCapture()
         while (streamPredicate == 0)
         {
             pthread_cond_wait(&cv, &condMutex);
-            StartCapture(1.0 / Streamer->getTargetFPS());
+            StartIntegration(1.0 / Streamer->getTargetFPS());
         }
 
         if (terminateThread)
@@ -371,11 +356,11 @@ void * RadioSim::streamCapture()
         s2 = ((double)tframe2.it_value.tv_sec) + ((double)tframe2.it_value.tv_usec / 1e6);
         deltas = fabs(s2 - s1);
 
-        if (deltas < CaptureTime)
-            usleep(fabs(CaptureTime - deltas) * 1e6);
+        if (deltas < IntegrationTime)
+            usleep(fabs(IntegrationTime - deltas) * 1e6);
 
-        uint32_t size = PrimaryDetector.getContinuumBufferSize();
-        Streamer->newFrame(PrimaryDetector.getContinuumBuffer(), size);
+        uint32_t size = getBufferSize();
+        Streamer->newFrame(getBuffer(), size);
 
         getitimer(ITIMER_REAL, &tframe2);
     }
